minor store refactor
